Removed "identity" from assurance level element. Made `id_verification_process` simply assurance_process
Following review, comments suggest that terms "identity_assurance_level" and "id_verification_process"  were too specific to the identity process and didn't really work for things like age verification, which is clearly in the spec. If they were made more general then the element is extendable to be used for other assurance processes that aren't specifically about identity, such as prove of age.